refactored condition and conditionEdit components into single component named condition
