+++ src/plugins/condition/plugin.js
@@ -43,9 +43,7 @@ export default function ConditionPlugin() {
             cssClass: 'icon-summary-widget',  // TODO: replace with class for new icon
             initialize: function (domainObject) {
                 domainObject.configuration = {
-                    conditionCollection: [
-                        
-                    ]
+                    conditionCollection: []
                 };
                 domainObject.composition = [];
             }
